Changed how plant positions work
Modified the calculations for tile offset.
Changed how plant prefabs work so that multiple sprites can be added to a plant prefab.
Added a capsule collider to the tree prefab, and partially polished functionality to change its size through various growth stages.
Fixed cell neighbour detection, it works now. Fixed cell positions on level 1 so the geometry is always 0 and the actual cell is what is being moved around.